# ND2: read chunkmap to improve performance