--- a/components/formats-gpl/src/loci/formats/in/NativeND2Reader.java
+++ b/components/formats-gpl/src/loci/formats/in/NativeND2Reader.java
@@ -696,7 +696,7 @@ public class NativeND2Reader extends FormatReader {
 
             blockCount --; // one was already added by the outer blockCount ++;
 
-            in.seek(lastImage.position + 16);
+            in.seek(lastImage.position + 16 + lastImage.length);
 
             continue;