See https://trello.com/c/Y4KMCatE/16-nd2-chunkmap-speedup-pr

This builds upon https://github.com/openmicroscopy/bioformats/pull/2556 from @csachs to fix remaining test issues and allow chunkmap processing to be disabled via the ```MetadataOptions``` API.

I would expect all tests to pass, with the subset/full repo jobs being slightly faster with this PR included.  

Expected offsets to image data within a file should now be logged at ```DEBUG```, regardless of whether or not chunkmap processing is enabled.  This can be tested with ```showinf -nopix -debug nd2/dallas/150609_m597_ut8.nd2``` and ```showinf -nopix -debug -option nativend2.chunkmap false nd2/dallas/150609_m597_ut8.nd2```.  ```setId``` times can also be compared in the same test, as the ```dallas``` file is quite large and should benefit from the performance improvement here.

With a build of ```bioformats_package.jar``` containing this PR installed in ImageJ, the ```Nikon ND2``` entry in ```Plugins > Bio-Formats > Bio-Formats Plugins Configuration``` should show a checkbox for using the chunkmap.  It should be checked by default, but it would be worth verifying that one or two files open identically with and without the box checked.