@sbesson 2017-03-07
Launched a [repository job](https://ci.openmicroscopy.org/view/Failing/job/BIOFORMATS-DEV-breaking-repository-subset/190) against the current `nd2` folder without this PR included to have a baseline. Labelling as `breaking` to test the impact of this change in isolation before moving into the daily builds.