@sbesson 2017-03-08
Tested using the build with this PR included versus the latest version of Bio-Formats (5.3.4). For the two of the largest ND2 samples in our repository:

````
[sbesson@ome-c6100-3 ~]$ for i in $(seq 1 3); do (./bftools-5.3.4/showinf -nopix /ome/data_repo/curated/nd2/dallas/150609_m597_ut8.nd2 | grep Init); done
Initializing reader
Initialization took 7.651s
Initializing reader
Initialization took 8.127s
Initializing reader
Initialization took 7.911s
[sbesson@ome-c6100-3 ~]$ for i in $(seq 1 3); do (/opt/hudson/workspace/BIOFORMATS-DEV-breaking-build/tools/showinf -nopix /ome/data_repo/curated/nd2/dallas/150609_m597_ut8.nd2 | grep Init); done
Initializing reader
Initialization took 2.69s
Initializing reader
Initialization took 2.233s
Initializing reader
Initialization took 2.493s
[sbesson@ome-c6100-3 ~]$ for i in $(seq 1 3); do (./bftools-5.3.4/showinf -nopix /ome/data_repo/curated/nd2/sara/Ex9\ ALADIN\ WT\ KO\ KI\ 31_07_2013.nd2 | grep Init); done
Initializing reader
Initialization took 12.967s
Initializing reader
Initialization took 12.745s
Initializing reader
Initialization took 13.851s
[sbesson@ome-c6100-3 ~]$ for i in $(seq 1 3); do (/opt/hudson/workspace/BIOFORMATS-DEV-breaking-build/tools/showinf -nopix /ome/data_repo/curated/nd2/sara/Ex9\ ALADIN\ WT\ KO\ KI\ 31_07_2013.nd2 | grep Init); done
Initializing reader
Initialization took 4.349s
Initializing reader
Initialization took 2.658s
Initializing reader
Initialization took 3.447s
````

So from the performance standpoint, this PR definitely seems to reduce the initialization time by a factor three. This also means we might want to reconsider enabling the `sara` dataset into the daily repository job.

All automated tests are passing so this PR is introducing no regression in the core metadata at the very minimum. Looking at the [time trends](https://ci.openmicroscopy.org/view/Failing/job/BIOFORMATS-DEV-breaking-repository-subset/buildTimeTrend), it is interesting to see the time with this PR included (builds 192, 195, 197) is larger than the builds without this PR (builds 193, 194). Could it be that while these changes increase performance for large ND2 files by scanning the chunkmap, the new code adds some overhead which slightly decreases performance for small ND2 files? Is there some additional profiling that can be done here?