@melissalinkert 2017-03-08
I think d8b5a97 will help with the discrepancy in build times - it certainly makes a big difference in a handful of files tried locally.  https://ci.openmicroscopy.org/view/Breaking/job/BIOFORMATS-DEV-breaking-repository-subset/200/ started to re-test.