@melissalinkert 2017-03-08
And now https://ci.openmicroscopy.org/view/Breaking/job/BIOFORMATS-DEV-breaking-repository-subset/201/ started to include 01bf93e and fix the one regressed file from build 200.