@sbesson 2017-03-08
With the latest set of commits, builds [201](https://ci.openmicroscopy.org/view/Breaking/job/BIOFORMATS-DEV-breaking-repository-subset/201), [202](https://ci.openmicroscopy.org/view/Breaking/job/BIOFORMATS-DEV-breaking-repository-subset/202) and [203](https://ci.openmicroscopy.org/view/Breaking/job/BIOFORMATS-DEV-breaking-repository-subset/203) all complete in 20min. Additionally, the initialization of the `sara/Ex9\ ALADIN\ WT\ KO\ KI\ 31_07_2013.nd2` file is now down to 1.8s. Very nice! Removing the `breaking` label for inclusion in the daily build.