@sbesson 2017-03-09
Also retested the impact of FS caching using a 50GB nd2 file stored on an external USB driver. Ran 6 successive opening of the files right after unmounting and remounting the USB hard drive.

Without this PR:

```
sbesson@c010156:Downloads $ for i in $(seq 1 6); do (showinf -nopix /Volumes/Harris_Drive/Big\ ND2\ File/O1_GROWTH_lowcellDens.nd2 | grep Initialization); done
Initialization took 84.426s
Initialization took 6.471s
Initialization took 6.706s
Initialization took 6.54s
Initialization took 6.687s
Initialization took 6.61s
```

With this PR:

```
sbesson@c010156:Downloads $ for i in $(seq 1 6); do (./bftools/showinf -nopix /Volumes/Harris_Drive/Big\ ND2\ File/O1_GROWTH_lowcellDens.nd2 | grep Initialization); done
Initialization took 1.427s
Initialization took 1.291s
Initialization took 1.224s
Initialization took 1.245s
Initialization took 1.247s
Initialization took 1.075s
```

So for this file, the `setId` performance gain is either 60/1 for a cold cache or 6/1 for a warm cache.

Also tested the plugin functionality (see screenshot below) using ImageJ in debug mode to check the output of the file opening call.
 
<img width="894" alt="screen shot 2017-03-09 at 12 43 07" src="https://cloud.githubusercontent.com/assets/1355463/23753167/49a64490-04d0-11e7-98db-f4f520649826.png">

In addition to the validation by the [full repository job](https://ci.openmicroscopy.org/view/Failing/job/BIOFORMATS-DEV-merge-full-repository/), one final check might be to review this PR via ImageJ/Fiji on Windows. Once cleared, this all looks great and ready to include for Bio-Formats 5.4.0.