@bramalingam 2017-03-10
Did the testing on imageJ, 

1) Checked the config options and can see the expected option (selected).
2) Toggling the checkbox and importing nd2 images did not alter the pixel/metadata on the few files I tried to import.

Looks good in my perspective.