@sbesson 2017-03-10
Thanks @bramalingam. As shown by https://ci.openmicroscopy.org/job/BIOFORMATS-DEV-merge-full-repository/465/, all repository tests are passing with this PR included and the `sara` dataset (largest ND2 file in our repository ~147G) is now tested daily as well.

Merging for inclusion in the upcoming Bio-Formats 5.4.0 release and deleting the `nd2_performance` integration branch which is now fully merged into develop. Thanks @csachs for your perseverance and your contribution.