@csachs 2017-03-10
Thanks for merging this.
The speed-ups were even more drastic at the time the work was started, and still are for larger files and slower links (eg > 100 GiB over 100 Mbit network).
Really happy this finally made it into develop. I'm sure many ND2 users will benefit from it :smiley: